datatypes converted to float
git-svn-id: svn://localhost:8080/caleydo_main/trunk@3304 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3

+++ org.caleydo.view.datawindows/src/org/caleydo/view/datawindows/DataWindowsMouseWheelListener.java
@@ -8,11 +8,11 @@ import java.awt.event.MouseWheelListener;
 public class DataWindowsMouseWheelListener extends MouseAdapter implements
 		MouseWheelListener, MouseMotionListener {
 	private GLHyperbolic hyperbolic;
-	private double wheelFactor;
+	private float wheelFactor;
 
 	public DataWindowsMouseWheelListener(GLHyperbolic master) {
 		hyperbolic = master;
-		wheelFactor = 0.2;
+		wheelFactor = 0.2f;
 	}
 
 	public void mouseWheelMoved(MouseWheelEvent event) {
